// src/common/dtos/query.pagination.dto.ts
import { IsNumberString, IsOptional } from 'class-validator'

export class QueryPaginationDTO {
  @IsOptional()
  @IsNumberString()
  page?: string

  @IsOptional()
  @IsNumberString()
  limit?: string
}

export interface PaginatedResponseDTO<T> {
  data: T[]
  meta: {
    total: number
    lastPage: number
    currentPage: number
    totalPerPage: number
    prevPage: number | null
    nextPage: number | null
  }
}